Stable Version
Add files with tested methods to import a text file with tags.

Also I updated the posts.sql script, which had incorrect usage of primary key settings.

Last but not least, the scrap function also gets the hashtag being scrapped and persists on the posts table, to help on further analysis.