Discovered a bug.
When working with session, there is a limit of 4093 bytes of data which I can use to store the post keys. Therefore, when the input hashtag is something popular, there will be more than 4093 (4KiB) data to keep on session and share between pages (posts and insights). Thus I need to think about some different way to pass data between pages other than session or cookies.